fix: AU should not update to pre-release versions
Partially fixes mkevenaar/chocolatey-packages#55

Windows Terminal is scheduled to get a new release around July 20. At that time, AU won't update the chocolatey package if the new release is a pre-release. If the release is a normal stable release, AU should update the chocolatey package.

PS: I have no experience in building chocolatey packages so I'm not 100% sure if this will work.