(intel-dsa) Migrates Package from BCurran3
This brings the package code across from BC's repository, and updates it to fit this ownership and repository style.

The update script could probably be trimmed back a little to just contain one of the two methods - parsing the release table is a bit brash when we could just download the binary. Hopefully it's acceptable - and/or it can be trimmed to the repository preference.